requirements.txt
pytest==6.2.4
pytest-env==0.6.2
pytest-cov==2.12.1
pytest-django==4.3.0
pytest-pythonpath==0.7.3
pytest-tldr==0.2.4
pytest-xdist==2.2.1
